feat: define maybeLogin decorator
